Add skeleton work for HTMLAudioElement
This currently just contains a lot of mostly-working auto-generated stubs for the HTMLMediaElement interface. Recent upgrades to webidl-class-generator and webidl-conversions have enabled that to generate a lot better, but it still doesn't *do* anything. We need to implement the delegation of the generated class to the implementation class next, and this gives us a playground in which to do it.